refactor: rename file to call-localize-compiler
